requirements.txt
Django==1.9.1
camel-snake-kebab==0.3.2
click==6.2
dj-database-url==0.3.0
django-jsonfield==0.9.15
djangorestframework==3.3.2
djangorestframework-camel-case==0.2.0
gunicorn==19.4.5
ordereddict==1.1
psycopg2==2.6.1
python-dotenv==0.3.0
requests==2.9.1
wsgiref==0.1.2
